Implementação do padrao Interpreter

+++ Design Patterns C#/Design Patterns/Interpreter/IExpressao.cs
@@ -2,6 +2,6 @@
 {
     public interface IExpressao
     {
-        int Avalia();
+        decimal Avalia();
     }
 }
